+++ sit-rdg-core/src/main/java/io/sitoolkit/rdg/core/domain/generator/config/SchemaConfig.java
@@ -1,13 +1,11 @@
 package io.sitoolkit.rdg.core.domain.generator.config;
 
-import java.util.Objects;
-import java.util.SortedSet;
-import java.util.TreeSet;
-
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
+import java.util.Objects;
+import java.util.SortedSet;
+import java.util.TreeSet;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
@@ -21,7 +19,7 @@ public class SchemaConfig {
   private String name;
 
   @JsonProperty("rowCount")
-  private Integer rowCount;
+  private Long rowCount;
 
   @JsonProperty("requiredValueCount")
   private Integer requiredValueCount;
@@ -30,7 +28,7 @@ public class SchemaConfig {
   @JsonProperty("tableConfigs")
   private SortedSet<TableConfig> tableConfigs = new TreeSet<>();
 
-  public Integer getRowCount() {
+  public Long getRowCount() {
     if (Objects.isNull(rowCount)) {
       rowCount = setting.getDefaultRowCount();
     }
